requirements.txt
Flask
psycopg2-binary
python-dotenv
Werkzeug
gunicorn
pytz